// src/app/core/model/security/passport-menu.model.ts
export class PassportMenuModel {
    id: number;
    icon: string;
    title: string;
    type: string;
    url: string;

    constructor() {}
}

export class PassportMenuOperacionModel {
    CODIGO: string;
    ID_MENU: number;
    ID_MENU_PADRE: number;
    NOMBRE_ICONO: string;
    NOMBRE_MENU: string;
    ORDEN_MENU: number;
    TIPO_OPCION: number;
    URL: string;
    constructor() {}
}

// export class PassportMenuOperacionModel {
//     CODIGO: string;
//     ID_MENU: number;
//     ID_MENU_PADRE: number;
//     NOMBRE_ICONO: string;
//     NOMBRE_MENU: string;
//     ORDEN_MENU: number;
//     TIPO_OPCION: number;
//     URL: string;
//     constructor() {
//     }
// }